+++ rust/Cargo.toml
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 [workspace]
-members = ["compactor", "sleeper_df", "rust_sketch"]
+members = [ "aggregator_udfs", "filter_udfs", "compactor", "rust_sketch", "sleeper_df" ]
 resolver = "2"
 
 [workspace.package]
@@ -32,7 +32,7 @@ aws-config = { version = "1.8.3" }                  # Credential loading
 aws-credential-types = { version = "1.2.4" }        # Credential provider types
 aws-types = { version = "1.3.8" }                   # for Region
 bytes = { version = "1.10.1" }                      # Byte buffer for S3 uploading
-cargo_metadata = { version = "0.20.0" }             # Access cargo metadata programmtically
+cargo_metadata = { version = "0.21.0" }             # Access cargo metadata programmtically
 chrono = { version = "0.4.41" }                     # Log helper
 clap = { version = "4.5.42" }                       # Cmd line args processing
 color-eyre = { version = "0.6.5" }                  # Colourised version of `anyhow`
