+++ rust/aggregator_udfs/Cargo.toml
@@ -1,0 +1,31 @@
+# Copyright 2022-2025 Crown Copyright
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+[package]
+name = "aggregator_udfs"
+description = "Apache DataFusion aggregation UDFs for Sleeper."
+keywords = ["sleeper"]
+categories = ["science"]
+version = { workspace = true }
+edition = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+authors = { workspace = true }
+rust-version = { workspace = true }
+publish = { workspace = true }
+
+[dependencies]
+arrow = { workspace = true }
+datafusion = { workspace = true }
+nohash = { workspace = true }
+num-traits = { workspace = true }
