+++ rust/aggregator_udfs/src/map_aggregate/accumulator.rs
