+++ rust/aggregator_udfs/src/map_aggregate/accumulator/string_accumulator.rs
@@ -1,4 +1,4 @@
-/// Simple [`Accumulator`] implementations for byte map aggregation.
+/// Simple [`Accumulator`] implementations for string map aggregation.
 /*
 * Copyright 2022-2025 Crown Copyright
 *
@@ -16,8 +16,8 @@
 */
 use arrow::{
     array::{
-        ArrayBuilder, ArrayRef, ArrowPrimitiveType, AsArray, BinaryBuilder, MapBuilder,
-        MapFieldNames, StructArray,
+        ArrayBuilder, ArrayRef, ArrowPrimitiveType, AsArray, MapBuilder, MapFieldNames,
+        StringBuilder, StructArray,
     },
     datatypes::DataType,
 };
@@ -30,19 +30,15 @@ use datafusion::{
 use num_traits::NumAssign;
 use std::{fmt::Debug, sync::Arc};
 
-use crate::datafusion::aggregate_udf::{
-    MapAggregatorOp, map_aggregate::aggregator::PrimBuilderType,
-};
-
 /// Given an Arrow [`StructArray`] of keys and values, update the given map.
 ///
-/// This implementation is for maps with byte keys and primitive values.
+/// This implementation is for maps with string keys and primitive values.
 ///
 /// All nulls keys/values are skipped over.
-fn update_byte_map<VBuilder>(
+fn update_string_map<VBuilder>(
     input: Option<&StructArray>,
     map: &mut HashMap<
-        Vec<u8>,
+        String,
         <<VBuilder as PrimBuilderType>::ArrowType as ArrowPrimitiveType>::Native,
     >,
     op: &MapAggregatorOp,
@@ -51,7 +47,7 @@ fn update_byte_map<VBuilder>(
     <<VBuilder as PrimBuilderType>::ArrowType as ArrowPrimitiveType>::Native: NumAssign + Ord,
 {
     if let Some(entries) = input {
-        let col1 = entries.column(0).as_binary::<i32>();
+        let col1 = entries.column(0).as_string::<i32>();
         let col2 = entries
             .column(1)
             .as_primitive::<<VBuilder as PrimBuilderType>::ArrowType>();
@@ -70,17 +66,17 @@ fn update_byte_map<VBuilder>(
 
 /// Single value primitive accumulator function for maps.
 #[derive(Debug)]
-pub struct ByteMapAccumulator<VBuilder>
+pub struct StringMapAccumulator<VBuilder>
 where
     VBuilder: ArrayBuilder + PrimBuilderType,
 {
     inner_field_type: DataType,
     values:
-        HashMap<Vec<u8>, <<VBuilder as PrimBuilderType>::ArrowType as ArrowPrimitiveType>::Native>,
+        HashMap<String, <<VBuilder as PrimBuilderType>::ArrowType as ArrowPrimitiveType>::Native>,
     op: MapAggregatorOp,
 }
 
-impl<VBuilder> ByteMapAccumulator<VBuilder>
+impl<VBuilder> StringMapAccumulator<VBuilder>
 where
     VBuilder: ArrayBuilder + PrimBuilderType,
 {
@@ -96,7 +92,7 @@ where
         if let DataType::Map(field, _) = map_type {
             let DataType::Struct(_) = field.data_type() else {
                 return plan_err!(
-                    "ByteMapAccumulator inner field type should be a DataType::Struct"
+                    "StringMapAccumulator inner field type should be a DataType::Struct"
                 );
             };
             Ok(Self {
@@ -105,7 +101,7 @@ where
                 op,
             })
         } else {
-            plan_err!("Invalid datatype for ByteMapAccumulator {map_type:?}")
+            plan_err!("Invalid datatype for StringMapAccumulator {map_type:?}")
         }
     }
 
@@ -116,7 +112,7 @@ where
     /// # Panics
     /// If an invalid map type is found. This condition shouldn't occur as it is checked
     /// upon construction.
-    fn make_map_builder(&self, cap: usize) -> MapBuilder<BinaryBuilder, VBuilder> {
+    fn make_map_builder(&self, cap: usize) -> MapBuilder<StringBuilder, VBuilder> {
         match &self.inner_field_type {
             DataType::Struct(fields) => {
                 let names = MapFieldNames {
@@ -124,34 +120,34 @@ where
                     value: fields[1].name().clone(),
                     entry: "key_value".into(),
                 };
-                let key_builder = BinaryBuilder::with_capacity(cap, 1024);
+                let key_builder = StringBuilder::with_capacity(cap, 1024);
                 let value_builder = VBuilder::default();
                 MapBuilder::with_capacity(Some(names), key_builder, value_builder, cap)
                     .with_keys_field(fields[0].clone())
                     .with_values_field(fields[1].clone())
             }
             _ => unreachable!(
-                "Invalid datatype inside ByteMapAccumulator {:?}",
+                "Invalid datatype inside StringMapAccumulator {:?}",
                 self.inner_field_type
             ),
         }
     }
 }
 
-impl<VBuilder> Accumulator for ByteMapAccumulator<VBuilder>
+impl<VBuilder> Accumulator for StringMapAccumulator<VBuilder>
 where
     VBuilder: ArrayBuilder + PrimBuilderType,
     <<VBuilder as PrimBuilderType>::ArrowType as ArrowPrimitiveType>::Native: NumAssign + Ord,
 {
     fn update_batch(&mut self, values: &[ArrayRef]) -> Result<()> {
         if values.len() != 1 {
-            return exec_err!("ByteMapAccumulator only accepts single column input");
+            return exec_err!("StringMapAccumulator only accepts single column input");
         }
 
         let input = values[0].as_map();
         // For each map we get, feed it into our internal aggregated map
         for map in input.iter() {
-            update_byte_map::<VBuilder>(map.as_ref(), &mut self.values, &self.op);
+            update_string_map::<VBuilder>(map.as_ref(), &mut self.values, &self.op);
         }
         Ok(())
     }
@@ -186,14 +182,14 @@ mod tests {
         datafusion::aggregate_udf::{
             MapAggregatorOp,
             map_aggregate::{
-                accumulator::byte_accumulator::update_byte_map,
+                accumulator::string_accumulator::{StringMapAccumulator, update_string_map},
                 aggregator::map_test_common::make_map_datatype,
             },
         },
     };
     use arrow::{
         array::{
-            AsArray, BinaryArray, BinaryBuilder, Int64Array, Int64Builder, StructBuilder,
+            AsArray, Int64Array, Int64Builder, StringArray, StringBuilder, StructBuilder,
             UInt16Builder,
         },
         datatypes::{DataType, Field, Fields, Int64Type},
@@ -203,51 +199,49 @@ mod tests {
     };
     use std::sync::Arc;
 
-    use super::ByteMapAccumulator;
-
-    // Macro to convert a literal to owned bytes
+    // Macro to stringify a literal
     macro_rules! s {
         ($l:literal) => {
-            String::from($l).into_bytes()
+            String::from($l)
         };
     }
 
     #[test]
-    fn update_byte_map_none() {
+    fn update_string_map_none() {
         // Given
-        let mut map = HashMap::<Vec<u8>, i64>::new();
+        let mut map = HashMap::<String, i64>::new();
         map.insert(s!["test"], 2);
         map.insert(s!["test2"], 4);
         let expected = map.clone();
 
         // When
-        update_byte_map::<Int64Builder>(None, &mut map, &MapAggregatorOp::Sum);
+        update_string_map::<Int64Builder>(None, &mut map, &MapAggregatorOp::Sum);
 
         // Then - expect no changes
         assert_eq!(map, expected);
     }
 
     #[test]
-    fn update_byte_map_empty_values() {
+    fn update_string_map_empty_values() {
         // Given
-        let mut map = HashMap::<Vec<u8>, i64>::new();
+        let mut map = HashMap::<String, i64>::new();
         map.insert(s!["test"], 2);
         map.insert(s!["test2"], 4);
         let expected = map.clone();
 
         let mut entry_builder = StructBuilder::new(
             Fields::from(vec![
-                Field::new("key", DataType::Binary, false),
+                Field::new("key", DataType::Utf8, false),
                 Field::new("value", DataType::Int64, false),
             ]),
             vec![
-                Box::new(BinaryBuilder::new()),
+                Box::new(StringBuilder::new()),
                 Box::new(Int64Builder::new()),
             ],
         );
 
         // When
-        update_byte_map::<Int64Builder>(
+        update_string_map::<Int64Builder>(
             Some(&entry_builder.finish()),
             &mut map,
             &MapAggregatorOp::Sum,
@@ -258,9 +252,9 @@ mod tests {
     }
 
     #[test]
-    fn update_byte_map_enters_first_values() {
+    fn update_string_map_enters_first_values() {
         // Given
-        let mut map = HashMap::<Vec<u8>, i64>::new();
+        let mut map = HashMap::<String, i64>::new();
         let mut expected = HashMap::new();
         expected.insert(s!["1"], 2);
         expected.insert(s!["3"], 4);
@@ -268,17 +262,17 @@ mod tests {
 
         let mut entry_builder = StructBuilder::new(
             Fields::from(vec![
-                Field::new("key", DataType::Binary, false),
+                Field::new("key", DataType::Utf8, false),
                 Field::new("value", DataType::Int64, false),
             ]),
             vec![
-                Box::new(BinaryBuilder::new()),
+                Box::new(StringBuilder::new()),
                 Box::new(Int64Builder::new()),
             ],
         );
         for (k, v) in [(s!["1"], 2), (s!["3"], 4), (s!["8"], -10)] {
             entry_builder
-                .field_builder::<BinaryBuilder>(0)
+                .field_builder::<StringBuilder>(0)
                 .unwrap()
                 .append_value(k);
             entry_builder
@@ -289,7 +283,7 @@ mod tests {
         }
 
         // When
-        update_byte_map::<Int64Builder>(
+        update_string_map::<Int64Builder>(
             Some(&entry_builder.finish()),
             &mut map,
             &MapAggregatorOp::Sum,
@@ -300,9 +294,9 @@ mod tests {
     }
 
     #[test]
-    fn update_byte_map_enters_values_and_sums() {
+    fn update_string_map_enters_values_and_sums() {
         // Given
-        let mut map = HashMap::<Vec<u8>, i64>::new();
+        let mut map = HashMap::<String, i64>::new();
         map.insert(s!["1"], 2);
         map.insert(s!["3"], 4);
         let mut expected = HashMap::new();
@@ -312,17 +306,17 @@ mod tests {
 
         let mut entry_builder = StructBuilder::new(
             Fields::from(vec![
-                Field::new("key", DataType::Binary, false),
+                Field::new("key", DataType::Utf8, false),
                 Field::new("value", DataType::Int64, false),
             ]),
             vec![
-                Box::new(BinaryBuilder::new()),
+                Box::new(StringBuilder::new()),
                 Box::new(Int64Builder::new()),
             ],
         );
         for (k, v) in [(s!["1"], 2), (s!["1"], 2), (s!["3"], 4), (s!["4"], 2)] {
             entry_builder
-                .field_builder::<BinaryBuilder>(0)
+                .field_builder::<StringBuilder>(0)
                 .unwrap()
                 .append_value(k);
             entry_builder
@@ -333,7 +327,7 @@ mod tests {
         }
 
         // When
-        update_byte_map::<Int64Builder>(
+        update_string_map::<Int64Builder>(
             Some(&entry_builder.finish()),
             &mut map,
             &MapAggregatorOp::Sum,
@@ -345,23 +339,23 @@ mod tests {
 
     #[test]
     #[should_panic(expected = "Nullable entries aren't supported")]
-    fn update_byte_map_panics_on_null() {
+    fn update_string_map_panics_on_null() {
         // Given
-        let mut map = HashMap::<Vec<u8>, i64>::new();
+        let mut map = HashMap::<String, i64>::new();
 
         let mut entry_builder = StructBuilder::new(
             Fields::from(vec![
-                Field::new("key", DataType::Binary, true),
+                Field::new("key", DataType::Utf8, true),
                 Field::new("value", DataType::Int64, true),
             ]),
             vec![
-                Box::new(BinaryBuilder::new()),
+                Box::new(StringBuilder::new()),
                 Box::new(Int64Builder::new()),
             ],
         );
         for (k, v) in [(s!["1"], 2), (s!["1"], 2), (s!["3"], 4), (s!["4"], 2)] {
             entry_builder
-                .field_builder::<BinaryBuilder>(0)
+                .field_builder::<StringBuilder>(0)
                 .unwrap()
                 .append_value(k);
             entry_builder
@@ -372,7 +366,7 @@ mod tests {
         }
         // Add null elements
         entry_builder
-            .field_builder::<BinaryBuilder>(0)
+            .field_builder::<StringBuilder>(0)
             .unwrap()
             .append_null();
         entry_builder
@@ -382,7 +376,7 @@ mod tests {
         entry_builder.append(true);
 
         // When - panic
-        update_byte_map::<Int64Builder>(
+        update_string_map::<Int64Builder>(
             Some(&entry_builder.finish()),
             &mut map,
             &MapAggregatorOp::Sum,
@@ -392,10 +386,10 @@ mod tests {
     #[test]
     fn try_new_should_succeed() {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
 
         // When
-        let acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
+        let acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
 
         // Then
         assert!(acc.is_ok());
@@ -405,13 +399,13 @@ mod tests {
     fn try_new_should_error_on_non_map_type() {
         // Given
         let mt = DataType::Int16;
-        let acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
+        let acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
 
         // Then
         assert_error!(
             acc,
             DataFusionError::Plan,
-            "Invalid datatype for ByteMapAccumulator Int16"
+            "Invalid datatype for StringMapAccumulator Int16"
         );
     }
 
@@ -420,22 +414,22 @@ mod tests {
         // Given
         let mt = DataType::Map(Arc::new(Field::new("test", DataType::Int16, false)), false);
         let acc: Result<_, DataFusionError> =
-            ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
+            StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum);
 
         // Then
         assert_error!(
             acc,
             DataFusionError::Plan,
-            "ByteMapAccumulator inner field type should be a DataType::Struct"
+            "StringMapAccumulator inner field type should be a DataType::Struct"
         );
     }
 
     #[test]
-    #[should_panic(expected = "Invalid datatype inside ByteMapAccumulator Int16")]
+    #[should_panic(expected = "Invalid datatype inside StringMapAccumulator Int16")]
     fn make_map_builder_check_unreachable() {
         // Given
         // Build instance directly via private constructor
-        let acc = ByteMapAccumulator::<Int64Builder> {
+        let acc = StringMapAccumulator::<Int64Builder> {
             inner_field_type: DataType::Int16,
             values: HashMap::new(),
             op: MapAggregatorOp::Sum,
@@ -448,9 +442,9 @@ mod tests {
     #[test]
     fn make_map_builder_field_names_equal() {
         // Given
-        let acc = ByteMapAccumulator::<Int64Builder> {
+        let acc = StringMapAccumulator::<Int64Builder> {
             inner_field_type: DataType::Struct(Fields::from(vec![
-                Field::new("key1_name", DataType::Binary, false),
+                Field::new("key1_name", DataType::Utf8, false),
                 Field::new("value1_name", DataType::Int64, false),
             ])),
             values: HashMap::new(),
@@ -470,9 +464,9 @@ mod tests {
     #[test]
     fn make_map_builder_field_types_equal() {
         // Given
-        let acc = ByteMapAccumulator::<UInt16Builder> {
+        let acc = StringMapAccumulator::<UInt16Builder> {
             inner_field_type: DataType::Struct(Fields::from(vec![
-                Field::new("key1_name", DataType::Binary, false),
+                Field::new("key1_name", DataType::Utf8, false),
                 Field::new("value1_name", DataType::UInt16, false),
             ])),
             values: HashMap::new(),
@@ -483,15 +477,15 @@ mod tests {
         let array = acc.make_map_builder(10).finish();
 
         // Then
-        assert_eq!(*array.key_type(), DataType::Binary);
+        assert_eq!(*array.key_type(), DataType::Utf8);
         assert_eq!(*array.value_type(), DataType::UInt16);
     }
 
     #[test]
     fn update_batch_should_fail_on_zero_or_multi_column() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
 
         // When
         let result = acc.update_batch(&[]);
@@ -500,11 +494,11 @@ mod tests {
         assert_error!(
             result,
             DataFusionError::Execution,
-            "ByteMapAccumulator only accepts single column input"
+            "StringMapAccumulator only accepts single column input"
         );
 
         // When
-        let arr1 = BinaryArray::from_vec(vec![b"a"]);
+        let arr1 = StringArray::from(vec!["a"]);
         let arr2 = Int64Array::from_value(1, 1);
         let result = acc.update_batch(&[Arc::new(arr1), Arc::new(arr2)]);
 
@@ -512,7 +506,7 @@ mod tests {
         assert_error!(
             result,
             DataFusionError::Execution,
-            "ByteMapAccumulator only accepts single column input"
+            "StringMapAccumulator only accepts single column input"
         );
 
         Ok(())
@@ -521,8 +515,8 @@ mod tests {
     #[test]
     fn update_batch_should_not_update_zero_maps() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
 
         // When
@@ -538,8 +532,8 @@ mod tests {
     #[test]
     fn update_batch_should_not_update_single_empty_map() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
 
         // When
@@ -556,8 +550,8 @@ mod tests {
     #[test]
     fn update_batch_should_update_single_map() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
         let expected = HashMap::from([(s!["3"], 10)]);
 
@@ -577,8 +571,8 @@ mod tests {
     #[test]
     fn update_batch_should_update_two_single_map() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
         let expected = HashMap::from([(s!["3"], 10), (s!["2"], 5)]);
 
@@ -602,8 +596,8 @@ mod tests {
     #[test]
     fn update_batch_should_update_multiple_maps() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
         let expected = HashMap::from([(s!["3"], 10), (s!["1"], 13), (s!["2"], 7), (s!["4"], 10)]);
 
@@ -637,8 +631,8 @@ mod tests {
     #[test]
     fn evaluate_should_produce_results() -> Result<(), DataFusionError> {
         // Given
-        let mt = make_map_datatype(DataType::Binary, DataType::Int64);
-        let mut acc = ByteMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
+        let mt = make_map_datatype(DataType::Utf8, DataType::Int64);
+        let mut acc = StringMapAccumulator::<Int64Builder>::try_new(&mt, MapAggregatorOp::Sum)?;
         let mut builder = acc.make_map_builder(10);
         let expected = HashMap::from([(s!["3"], 10), (s!["1"], 9), (s!["2"], 7)]);
 
@@ -663,12 +657,12 @@ mod tests {
         let ScalarValue::Map(result_map) = result else {
             panic!("result type not a map");
         };
-        let keys = result_map.keys().as_binary::<i32>();
+        let keys = result_map.keys().as_string::<i32>();
         let values = result_map.values().as_primitive::<Int64Type>();
         let mut accumulated_map = HashMap::new();
         for item in keys.iter().zip(values.iter()) {
             if let (Some(opt_k), Some(opt_v)) = item {
-                accumulated_map.insert(opt_k.to_vec(), opt_v);
+                accumulated_map.insert(opt_k.to_string(), opt_v);
             }
         }
