+++ rust/aggregator_udfs/src/map_aggregate/aggregator.rs
@@ -32,20 +32,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-use crate::datafusion::aggregate_udf::{
-    MapAggregatorOp,
-    map_aggregate::{
-        accumulator::{
-            byte_accumulator::ByteMapAccumulator, prim_accumulator::PrimMapAccumulator,
-            string_accumulator::StringMapAccumulator,
-        },
-        group_accumulator::{
-            byte_group_accumulator::ByteGroupMapAccumulator,
-            prim_group_accumulator::PrimGroupMapAccumulator,
-            string_group_accumulator::StringGroupMapAccumulator,
-        },
-    },
-};
 use arrow::{
     array::{ArrowPrimitiveType, PrimitiveBuilder, downcast_integer},
     datatypes::{DataType, Fields},
