+++ rust/aggregator_udfs/src/map_aggregate/state.rs
