+++ rust/sleeper_df/Cargo.toml
@@ -39,15 +39,12 @@ aws-types = { workspace = true }
 bytes = { workspace = true }
 chrono = { workspace = true }
 color-eyre = { workspace = true }
-cxx = { workspace = true }
 datafusion = { workspace = true }
 env_logger = { workspace = true }
 futures = { workspace = true }
 libc = { workspace = true }
 log = { workspace = true }
-nohash = { workspace = true }
 num-format = { workspace = true }
-num-traits = { workspace = true }
 object_store = { workspace = true, features = [ "aws" ] }
 regex = { workspace = true }
 rust_sketch = { path = "../rust_sketch" }
